// lib/robotevents/robotevents/models/Event.ts
export interface Event {
  /**
   *
   * @type {number}
   * @memberof Event
   */
  id: number;
  /**
   *
   * @type {string}
   * @memberof Event
   */
  sku: string;
  /**
   *
   * @type {string}
   * @memberof Event
   */
  name: string;
  /**
   *
   * @type {Date}
   * @memberof Event
   */
  start?: Date;
  /**
   *
   * @type {Date}
   * @memberof Event
   */
  end?: Date;
  /**
   *
   * @type {IdInfo}
   * @memberof Event
   */
  season: IdInfo;
  /**
   *
   * @type {IdInfo}
   * @memberof Event
   */
  program: IdInfo;
  /**
   *
   * @type {Location}
   * @memberof Event
   */
  location: Location;
  /**
   *
   * @type {Array<{ [key: string]: Location; }>}
   * @memberof Event
   */
  locations?: Array<{ [key: string]: Location }>;
  /**
   *
   * @type {Array<Division>}
   * @memberof Event
   */
  divisions?: Array<Division>;
  /**
   *
   * @type {EventLevel}
   * @memberof Event
   */
  level?: EventLevel;
  /**
   *
   * @type {boolean}
   * @memberof Event
   */
  ongoing?: boolean;
  /**
   *
   * @type {boolean}
   * @memberof Event
   */
  awardsFinalized?: boolean;
  /**
   *
   * @type {EventType}
   * @memberof Event
   */
  eventType?: EventType;
}