// lib/robotevents/robotevents/models/Team.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public Robot Events API
 * An API to access data on Robot Events officially. ## Request Metadata Pagination is performed the same way throughout each pageable endpoint using query parameters `page` and `per_page` (per_page default is 25 with a maximum of 250). <br> Please note that dates should be formatted according to RFC3339. 
 *
 * The version of the OpenAPI document: 1.0.21
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Grade } from './Grade';
import {
    GradeFromJSON,
    GradeFromJSONTyped,
    GradeToJSON,
    GradeToJSONTyped,
} from './Grade';
import type { IdInfo } from './IdInfo';
import {
    IdInfoFromJSON,
    IdInfoFromJSONTyped,
    IdInfoToJSON,
    IdInfoToJSONTyped,
} from './IdInfo';
import type { Location } from './Location';
import {
    LocationFromJSON,
    LocationFromJSONTyped,
    LocationToJSON,
    LocationToJSONTyped,
} from './Location';

/**
 * 
 * @export
 * @interface Team
 */
export interface Team {
    /**
     * 
     * @type {number}
     * @memberof Team
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    number: string;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    teamName?: string;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    robotName?: string;
    /**
     * 
     * @type {string}
     * @memberof Team
     */
    organization?: string;
    /**
     * 
     * @type {Location}
     * @memberof Team
     */
    location?: Location;
    /**
     * 
     * @type {boolean}
     * @memberof Team
     */
    registered?: boolean;
    /**
     * 
     * @type {IdInfo}
     * @memberof Team
     */
    program: IdInfo;
    /**
     * 
     * @type {Grade}
     * @memberof Team
     */
    grade?: Grade;
}



/**
 * Check if a given object implements the Team interface.
 */
export function instanceOfTeam(value: object): value is Team {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('number' in value) || value['number'] === undefined) return false;
    if (!('program' in value) || value['program'] === undefined) return false;
    return true;
}

export function TeamFromJSON(json: any): Team {
    return TeamFromJSONTyped(json, false);
}

export function TeamFromJSONTyped(json: any, ignoreDiscriminator: boolean): Team {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'number': json['number'],
        'teamName': json['team_name'] == null ? undefined : json['team_name'],
        'robotName': json['robot_name'] == null ? undefined : json['robot_name'],
        'organization': json['organization'] == null ? undefined : json['organization'],
        'location': json['location'] == null ? undefined : LocationFromJSON(json['location']),
        'registered': json['registered'] == null ? undefined : json['registered'],
        'program': IdInfoFromJSON(json['program']),
        'grade': json['grade'] == null ? undefined : GradeFromJSON(json['grade']),
    };
}

export function TeamToJSON(json: any): Team {
    return TeamToJSONTyped(json, false);
}

export function TeamToJSONTyped(value?: Team | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'number': value['number'],
        'team_name': value['teamName'],
        'robot_name': value['robotName'],
        'organization': value['organization'],
        'location': LocationToJSON(value['location']),
        'registered': value['registered'],
        'program': IdInfoToJSON(value['program']),
        'grade': GradeToJSON(value['grade']),
    };
}